telephony_app/requirements.txt
vocode==0.1.109
redis
twilio
pyngrok
python-dotenv